[♻️refactor]: 페이지 파일에서 타입 분리
